class docs and moved some stuff around
