+++ addon/utils/array-methods.js
@@ -4,14 +4,36 @@ const {
   A
 } = Ember;
 
-// Splits an array and tacks the first chunk to the end of the second chunk
-export function arraySwap(octave, noteToShiftAt) {
-  const endOfOctave = octave.slice(0, noteToShiftAt);
-  const beginOfOctave = A(octave.slice(noteToShiftAt));
-  return beginOfOctave.pushObjects(endOfOctave);
+/**
+ * @module array-methods
+ * @private
+ * @ignore
+ */
+
+/**
+ * Given an array and an index, splits the array at index and pushes the first
+ * chunk to the end of the second chunk.
+ *
+ * @private
+ * @method arraySwap
+ * @param {array} arr An array to split, shift and rejoin.
+ * @param {number} index The index where the split should occur.
+ * @returns {array} The swapped/shifted array.
+ */
+export function arraySwap(arr, index) {
+  const endOfArr = arr.slice(0, index);
+  const beginOfArr = A(arr.slice(index));
+  return beginOfArr.pushObjects(endOfArr);
 }
 
-// Flattens an array of arrays into a single array
+/**
+ * Flattens an array of arrays into a shallow array.
+ *
+ * @private
+ * @method flatten
+ * @param {arrayOfArrays} arr An array to flatten.
+ * @returns {array} The flattened array.
+ */
 export function flatten(arrayOfArrays) {
   return A(arrayOfArrays).reduce((a, b) => A(a).concat(b));
 }
