+++ addon/utils/beat-track.js
@@ -6,6 +6,22 @@ const {
   computed
 } = Ember;
 
+/**
+ * An instance of this class has a single
+ * {@link https://developer.mozilla.org/en-US/docs/Web/API/AudioBuffer AudioBuffer}
+ * (a single "sound") but provides methods to play that sound repeatedly and in
+ * a rhythmic way. An instance of this class behaves very similarly to a "lane"
+ * on a drum machine or a step sequencer.
+ *
+ * <style>
+ *   .ignore-this--this-is-here-to-hide-constructor,
+ *   #BeatTrack { display: none; }
+ * </style>
+ *
+ * @see {@link https://developer.mozilla.org/en-US/docs/Web/API/AudioBuffer}
+ * @class BeatTrack
+ * @extends Sound
+ */
 const BeatTrack = Sound.extend({
   numBeats: 4,
   playingTime: 100,
