+++ addon/utils/note.js
@@ -1,30 +1,30 @@
 import Ember from 'ember';
-import { arraySwap, flatten } from './array-methods';
 import Sound from './sound';
 
 const {
-  A,
   computed
 } = Ember;
 
 /**
- * Ember.Object that acts as "note".
- * Set it's letter, octave, and (optionally) it's accidental.
+ * A class that represents a musical note. When a soundfont is created via the
+ * {@link module:Audio-Service Audio Service}, an instance of this class is
+ * created for every musical note that is represented in the soundfont. This
+ * class can be used for any collection of Sound instances where each instance
+ * needs an awareness of what "musical note" it is (i.e. octave, accidental, etc..).
  *
- * Computed Properties (examples for notes "Db1" and "D0" respectively):
+ * <style>
+ *   .ignore-this--this-is-here-to-hide-constructor,
+ *   #Note { display: none; }
+ * </style>
  *
- * get {string} identifier  examples: "Db1", "D0"
- * get {string} name        examples: "Db", "D"
+ * @class Note
+ * @extends Sound
  */
 export const Note = Sound.extend({
   letter: null,
   accidental: null,
   octave: null,
 
-  /**
-   * @computed identifier
-   * @return {string} computed note name including the octave
-   */
   identifier: computed('letter', 'accidental', 'octave', {
     get() {
       const accidental = this.get('accidental');
@@ -42,10 +42,6 @@ export const Note = Sound.extend({
     }
   }),
 
-  /**
-   * @computed name
-   * @return {string} computed note name, not including octave
-   */
   name: computed('letter', 'accidental', {
     get() {
       const accidental = this.get('accidental');
@@ -62,124 +58,3 @@ export const Note = Sound.extend({
     }
   })
 });
-
-/**
- * octaveShift - Takes an array of arrays of notes, determines the last note of
- * the first array, then splits the rest of the arrays in the array at the last
- * note of the first array, and moves the beginning of the array to the end
- * so that each array starts at the next note after the last note of the first
- * array, instead of at "A" (alphabetically)
- *
- * This is hard to explain. Here's an example.
- * (Simplified, as the real notes are objects.)
- *
- * Example input: [['A0', 'B0'], ['A1', 'B1', 'C1', 'D1']]
- * Example output: [['A0', 'B0'], ['C1', 'D1', 'A1', 'B1']]
- *
- * @param  {array} octaves  an array of octaves, each octave is an array of notes, each note is an Ember.Object
- * @return {array}          same array but shifted around
- */
-export function octaveShift(octaves) {
-  // Pull first octave from beginning of array
-  const firstOctave = A(A(octaves).shiftObject());
-  // Get all the note names from the second octave for comparison
-  const secondOctaveNames = A(octaves.get('firstObject')).getEach('name');
-  // Get the note name of the last note in the first octave
-  const lastNote = firstOctave.get('lastObject.name');
-  // Get the index of the occurrence of the last note from the first
-  // octave, in the second octave
-  const indexToShiftAt = secondOctaveNames.lastIndexOf(lastNote) + 1;
-  // Split the octave array at that point, and move the first chunk to the end
-  return A(octaves.map((octave) => arraySwap(octave, indexToShiftAt)))
-  // Put first octave back at the beginning of the array
-  .unshiftObjects([firstOctave]);
-}
-
-/**
- * octaveSort - Maps through an array of arrays and sorts each array with
- * "noteSort"
- *
- * @param  {array} octaves array of arrays to be sorted
- * @return {array}         array of sorted arrays
- */
-export function octaveSort(octaves) {
-  return octaves.map((octave) => octave.sort(noteSort));
-}
-
-/**
- * extractOctaves - Accepts an array of Note objects and passes back an array
- * like this: [original array, array of each octave in the orginal array]
- *
- * @param  {array} notes array of note objects
- * @return {array}       array containing two inner arrays, [0] is the untouched
- * input array, [1] is an array of all the octaves in the original array
- */
-export function extractOctaves(notes) {
-  return [ notes, A(A(notes).getEach('octave')) ];
-}
-
-/**
- * stripDuplicateOctaves - Accepts an array of two arrays and returns the same
- * array, but with array at index [1] uniq'd and sorted alphabetically
- *
- * @param  {array} [ notes, octaves ] the output from extractOctaves (above)
- * @return {array}
- */
-export function stripDuplicateOctaves([ notes, octaves ]) {
-  return [ notes, A(octaves).uniq().sort() ];
-}
-
-/**
- * createOctavesWithNotes - Accepts an array of two arrays, [0] being an array
- * of Note objects, [1] being all the available octaves. Returns a single array
- * made up of arrays of Note objects, organized by octave. Each inner array
- * represents all of the notes in an octave.
- *
- * @param  {array} data The output of stripDuplicateOctaves (above)
- * @return {array}
- */
-export function createOctavesWithNotes([ notes, octaves ]) {
-  return A(octaves).map((octave) => A(notes).filterBy('octave', octave));
-}
-
-/**
- * noteSort - Acts as a comparator function for javascript's array sort() method.
- * Sorts two Ember.Object "notes" alphabetically, flats before naturals.
- */
-export function noteSort(a, b) {
-  const aLet = a.get('letter');
-  const bLet = b.get('letter');
-
-  if (aLet < bLet) {
-    return -1;
-  }
-
-  if (aLet === bLet) {
-    if (a.get('accidental') === 'b') {
-      return -1;
-    }
-  }
-
-  return 1;
-}
-
-export function sortNotes(notes) {
-  // get octaves so that we can sort based on them
-  let sortedNotes = extractOctaves(notes);
-
-  // Each octave has tons of duplicates
-  sortedNotes = stripDuplicateOctaves(sortedNotes);
-
-  // Create array of arrays. Each inner array contains all the notes in an octave
-  sortedNotes = createOctavesWithNotes(sortedNotes);
-
-  // Sort the notes in each octave, alphabetically, flats before naturals
-  sortedNotes = octaveSort(sortedNotes);
-
-  // Determine last note of first octave, then for each octave, split at
-  // that note, then shift the beginning notes to the end
-  sortedNotes = octaveShift(sortedNotes);
-
-  // Flatten array of arrays into a flat array
-  return flatten(sortedNotes);
-}
