+++ app/utils/note-methods.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-audio/utils/note-methods';
