+++ tests/unit/utils/note-methods-test.js
@@ -1,0 +1,10 @@
+import noteMethods from 'dummy/utils/note-methods';
+import { module, test } from 'qunit';
+
+module('Unit | Utility | note methods');
+
+// Replace this with your real tests.
+test('it works', function(assert) {
+  let result = noteMethods();
+  assert.ok(result);
+});
